move module and rename package
